+++ components/stays/CategoryModal.jsx
