+++ components/stays/CategoryTool.jsx
