+++ components/stays/SearchOption.jsx
