+++ components/stays/SearchOptions.jsx
