+++ components/stays/StayPage.jsx
