+++ components/stays/TypeModal.jsx
