+++ components/stays/TypeTool.jsx
